revert ios changes

+++ Plugin.Maui.Apptentive/Platforms/iOS/Apptentive.ios.cs
@@ -138,6 +138,6 @@ partial class ApptentiveImplementation: IApptentive
 
     private void HandleAuthenticationFailed(ApptentiveAuthenticationFailureReason reason, string? error)
     {
-        AuthenticationFailed?.Invoke((long)reason, error);
+        AuthenticationFailed?.Invoke(reason, error);
     }
 }
